+++ src/Controlador/DetallecompraJpaController.java
@@ -104,7 +104,7 @@ public class DetallecompraJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Integer id = detallecompra.getIdDetalleCompra();
+                Long id = detallecompra.getIdDetalleCompra();
                 if (findDetallecompra(id) == null) {
                     throw new NonexistentEntityException("The detallecompra with id " + id + " no longer exists.");
                 }
@@ -117,7 +117,7 @@ public class DetallecompraJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Long id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -172,7 +172,7 @@ public class DetallecompraJpaController implements Serializable {
         }
     }
 
-    public Detallecompra findDetallecompra(Integer id) {
+    public Detallecompra findDetallecompra(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Detallecompra.class, id);
